debugged first version of execution start
